move in-memory verification store to module
