+++ core/common/service-types/build.gradle.kts
@@ -7,6 +7,7 @@ dependencies {
     annotationProcessor("org.immutables:value")
 
     api(project(":api-types"))
+    api(project(":core:common:api-types"))
     api(project(":core:data"))
     api("com.fasterxml.jackson.core:jackson-databind")
     api("org.immutables:value-annotations")
