+++ module/store/common/inmemory/src/main/java/org/example/age/common/service/store/inmemory/InMemoryVerificationStoreModule.java
@@ -1,0 +1,13 @@
+package org.example.age.common.service.store.inmemory;
+
+import dagger.Binds;
+import dagger.Module;
+import org.example.age.common.service.store.VerificationStore;
+
+/** Dagger module that publishes a binding for {@link VerificationStore}, which uses an in-memory store. */
+@Module
+public interface InMemoryVerificationStoreModule {
+
+    @Binds
+    VerificationStore bindVerificationStateStore(InMemoryVerificationStore impl);
+}
